Extract abstract resource module to toolkit common lib
